+++ js/main.js
@@ -1,3 +1,3 @@
-import {createPhoto} from './data';
+import {createPhoto} from './data_generator';
 
 export const photosList = Array.from({length: 25}, createPhoto);
